Collectables v1
The collectables will now be requested by the Game Manager, and selecting the right one will request for another one. When an object is selected it will stop spinning with the planet until another item is selected. Items appear within a time frame depending on if it is day or night.